src/main/java/com/foodrecord/core/disaster/IntelligentBackupManager.java
package com.foodrecord.core.disaster;

import org.springframework.stereotype.Component;

import java.util.Map;
import java.util.Set;
import java.util.concurrent.*;

@Component
public class IntelligentBackupManager extends BackupManager {
    private final DataChangeAnalyzer changeAnalyzer;
    private final BackupScheduleOptimizer scheduleOptimizer;
    private final StorageOptimizer storageOptimizer;
    
    public IntelligentBackupManager(
            DatabaseBackupService dbBackupService,
            FileSystemBackupService fsBackupService,
            BackupStorageService storageService,
            BackupVerificationService verificationService,
            DataChangeAnalyzer changeAnalyzer,
            BackupScheduleOptimizer scheduleOptimizer,
            StorageOptimizer storageOptimizer) {
        super(dbBackupService, fsBackupService, storageService, verificationService);
        this.changeAnalyzer = changeAnalyzer;
        this.scheduleOptimizer = scheduleOptimizer;
        this.storageOptimizer = storageOptimizer;
    }
    
    @Override
    public CompletableFuture<BackupResult> performFullBackup() {
        // 1. 分析数据变化
        ChangeAnalysis changes = changeAnalyzer.analyzeChanges();
        
        // 2. 确定备份策略
        BackupStrategy strategy = determineBackupStrategy(changes);
        
        // 3. 优化存储
        StorageOptimization storageOpt = storageOptimizer.optimize(changes);
        
        return CompletableFuture.supplyAsync(() -> {
            try {
                // 4. 执行智能备份
                BackupExecution execution = new BackupExecution(strategy, storageOpt);
                execution.setCompressionLevel(determineCompressionLevel(changes));
                execution.setParallelism(calculateOptimalParallelism());
                
                // 5. 执行备份
                BackupResult result = execution.execute();
                
                // 6. 更新备份统计
                updateBackupStatistics(result);
                
                // 7. 优化未来备份计划
                scheduleOptimizer.optimize(result);
                
                return result;
            } catch (Exception e) {
                handleBackupFailure(e);
                throw e;
            }
        });
    }
    
    private BackupStrategy determineBackupStrategy(ChangeAnalysis changes) {
        return BackupStrategy.builder()
            .type(selectBackupType(changes))
            .priority(calculatePriority(changes))
            .retentionPolicy(determineRetentionPolicy(changes))
            .verificationLevel(determineVerificationLevel(changes))
            .build();
    }

    private BackupType selectBackupType(ChangeAnalysis changes) {
        if (changes.getTotalChanges() > 1000 || changes.getImportance() == ImportanceLevel.CRITICAL) {
            return BackupType.FULL;
        }
        return BackupType.INCREMENTAL;
    }

    private int calculatePriority(ChangeAnalysis changes) {
        switch (changes.getImportance()) {
            case CRITICAL: return 1;
            case HIGH: return 2;
            case MEDIUM: return 3;
            default: return 4;
        }
    }

    private RetentionPolicy determineRetentionPolicy(ChangeAnalysis changes) {
        switch (changes.getImportance()) {
            case CRITICAL: return RetentionPolicy.PERMANENT;
            case HIGH: return RetentionPolicy.LONG_TERM;
            case MEDIUM: return RetentionPolicy.MEDIUM_TERM;
            default: return RetentionPolicy.SHORT_TERM;
        }
    }

    private VerificationLevel determineVerificationLevel(ChangeAnalysis changes) {
        switch (changes.getImportance()) {
            case CRITICAL: return VerificationLevel.COMPREHENSIVE;
            case HIGH: return VerificationLevel.THOROUGH;
            case MEDIUM: return VerificationLevel.STANDARD;
            default: return VerificationLevel.BASIC;
        }
    }

    private int determineCompressionLevel(ChangeAnalysis changes) {
        return changes.getImportance() == ImportanceLevel.CRITICAL ? 1 : 9;
    }

    private int calculateOptimalParallelism() {
        return Runtime.getRuntime().availableProcessors();
    }

    private void updateBackupStatistics(BackupResult result) {
        // 实现备份统计更新逻辑
    }

    private void handleBackupFailure(Exception e) {
        // 实现备份失败处理逻辑
    }

    public static class ChangeAnalysis {
        private final long totalChanges;
        private final Map<String, Long> changesByEntity;
        private final double changeRate;
        private final ImportanceLevel importance;
        private final Set<String> criticalChanges;

        public ChangeAnalysis(long totalChanges, Map<String, Long> changesByEntity, double changeRate, ImportanceLevel importance, Set<String> criticalChanges) {
            this.totalChanges = totalChanges;
            this.changesByEntity = changesByEntity;
            this.changeRate = changeRate;
            this.importance = importance;
            this.criticalChanges = criticalChanges;
        }

        public long getTotalChanges() {
            return totalChanges;
        }

        public Map<String, Long> getChangesByEntity() {
            return changesByEntity;
        }

        public double getChangeRate() {
            return changeRate;
        }

        public ImportanceLevel getImportance() {
            return importance;
        }

        public Set<String> getCriticalChanges() {
            return criticalChanges;
        }
    }
} 