src/test/java/com/foodrecord/core/disaster/IntelligentBackupManagerTest.java
package com.foodrecord.core.disaster;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import static org.mockito.Mockito.*;
import static org.assertj.core.api.Assertions.*;

@ExtendWith(MockitoExtension.class)
class IntelligentBackupManagerTest {

//    @Mock
//    private DataChangeAnalyzer changeAnalyzer;
//
//    @Mock
//    private BackupScheduleOptimizer scheduleOptimizer;
//
//    @Mock
//    private StorageOptimizer storageOptimizer;
//
//    @InjectMocks
//    private IntelligentBackupManager backupManager;
//
//    @Test
//    void shouldPerformBackupSuccessfully() {
//        // Given
//        ChangeAnalysis mockAnalysis = mock(ChangeAnalysis.class);
//        when(changeAnalyzer.analyzeChanges()).thenReturn(mockAnalysis);
//
//        StorageOptimization mockOptimization = mock(StorageOptimization.class);
//        when(storageOptimizer.optimize(any())).thenReturn(mockOptimization);
//
//        // When
//        CompletableFuture<BackupResult> future = backupManager.performBackup();
//        BackupResult result = future.join();
//
//        // Then
//        assertThat(result).isNotNull();
//        assertThat(result.isSuccess()).isTrue();
//        verify(scheduleOptimizer).optimize(any());
//    }
//
//    @Test
//    void shouldHandleBackupFailureGracefully() {
//        // Given
//        when(changeAnalyzer.analyzeChanges())
//            .thenThrow(new RuntimeException("Simulated failure"));
//
//        // When
//        CompletableFuture<BackupResult> future = backupManager.performBackup();
//
//        // Then
//        assertThatThrownBy(() -> future.join())
//            .hasCauseInstanceOf(RuntimeException.class)
//            .hasMessageContaining("Simulated failure");
//    }
} 